Wordle.jl Documentation

Overview

This module contains a New York Times Wordle solver. It is optimized to solve Wordle puzzles based on most frequently used 5 letter words. It provides the number of steps to solve along with the intermediate guesses.

There is an associated Wordle Jupyter notebook at src/WordleTest.ipynb.

Solver Strategy

Initial Conditions:

Set puzzle_word

X = Start with universe of 5 letter words along with frequency of usage.

Set current_universe = X

Start

Pick guess (by default use the function get_next_word).

If guess == puzzle_word

Goto End

Get wordle match info about how close guess is to the correct word:

wordle_info = create_wordle_info(<guess>, <puzzle_word>)

Example: wordle_info, create_wordle_info("exact", "crane") =

([('a', 3), ('e', -1), ('c', -4)], Dict('x' => (0, 0), 'c' => (1, 1), 'e' => (1, 1), 't' => (0, 0)))

Use this match info to filter existing universe of words.

current_universe = filter_universe(wordle_info, current_universe)

Goto Start

End

Return guess

Primary Function

Wordle.solve_wordle — Function

solve_wordle(puzzle_word, universe_df, rec_count, sol_path, last_guess,
                lfa; <keyword arguments>)

Solves a NYT Wordle puzzle.

By default, makes guesses based on function, get_next_word. However, there is an option to pass in a guessing strategy function.

Input Contract

The universe DataFrame is sorted from highest frequency to lowest.

universe_df schema is (:word, :freq). Define words, freq, and N by:

words = universe_df[:words];

freq = universe[:freq];

N = |universe|

∃ m > 0, ∀ i∈[1,N], |words[i]| == m $\quad$ (All the words in universe_df have the same length. )

words == words[sortperm(wts, rev=true)] $\quad$ (Words are sorted from highest to lowest by word usage.)

Arguments

puzzle_word::String : The puzzle word.

universe_df::DataFrame : A DataFrame with schema: word(words of the same length), freq(freq fraction by use)

NOTE: The universe is assumed to be sorted in reverse order by the :freq column.

rec_count::Int : The number of calls to this function.

sol_path::Vector{Any} : Any containing the current list of guesses: [ (guess, exact_info, universe_size) ...]

last_guess::String : The previous guess.

lfa::Vector{Char} : The lowercase alphabet listed in frequency-of-use order.

Keyword Arguments

chk_inputs::Bool : If true, check the input contract.

guess_strategy::Union{Function,Nothng} : If not nothing, apply this function to pick the next guess. If nothing, pick based on the function get_next_word.

ul::Int : The lower threshold size of the filtered Wordle universe.

uu::Int : The upper threshold size of the filtered Wordle universe.

init_guess::String : The starting guess to use.

Here,

exact_info has the form: [(LETTER, POSITION) ...]

universe_size is the size the word list when the guess was made.

The guess_strategy is only turned on when the filtered Wordle universe is between the thresholds: ul and uu; otherwise, the default strategy is used – the most frequently used word in the existing filtered Wordle universe is chosen.

Return

(sol_path, number-of-guesses, :SUCCESS/:FAILURE)

NOTE: A sol_path that does not include the puzzle word, means that at some point after a guess was made – along with the corresponding filtering of the universe – there was only one word left. In this case the guess count was increased by 1, but the function did not recurse.

Examples

julia> solve_wordle("taste"; init_guess="their")

(Any[(String7("their"), [('t', 1), ('e', -3)], 3585), 
     (String7("taken"), [('t', 1), ('a', 2), ('e', -4)], 34), 
	 (String7("table"), [('t', 1), ('a', 2), ('e', 5)], 3), 
	 (String7("taste"), [('t', 1), ('a', 2), ('s', 3), ('t', 4), ('e', 5)], 2)], 4, :SUCCESS)

Supporting Functions

Wordle.create_wordle_info — Function

create_wordle_info(guess, pword)

Create an information structure of the form: ([LETTER, EXACT_MATCH_POSITION)], Dict(LETTER => (NUMBER_OF_MATCHES, MATCH_FLAG))

Here, the dictionary has the inexact match information: of words.

LETTER : A matching letter

EXACT_MATCH_POSITION: The position (1-based index) if an exact match; OR minus the position if the letter is used, but not at this position.

NUMBER_OF_MATCHES : The number of matches.

The latter is interpreted thusly:

If MATCH_FLAG is 0: There are exactly NUMBER_OF_MATCHES with this letter that should occur in the puzzle word in addition to any existing exact matches of this letter.

Else (flag is 1) : There are at least NUMBER_OF_MATCHES with this letter that should occur in the puzzle word in addition to any existing exact matches of this letter.

Type Constraints

T <: AbstractString

Arguments

guess::T: The guess for the puzzle.

pword::T: The puzzle word.

Return

A tuple of a vector of tuples of exact matches and a dictionary of inexact match info.

Examples

julia> create_wordle_info("which", "where")
([('w', 1), ('h', 2)], Dict('h' => (0, 0), 'c' => (0, 0), 'i' => (0, 0)))

julia> create_wordle_info("teens", "where")
([('e', 3), ('e', -2)], Dict('n' => (0, 0), 's' => (0, 0), 't' => (0, 0), 'e' => (1, 1)))

Wordle.filter_universe — Function

filter_universe(wordle_info, words)

Filter an existing universe of words based on match info.

Type Constraints

T <: AbstractString

Arguments

wordle_info : Wordle info of the form: ([(LETTER, EXACT_POSITION)], Dict( LETTER => (k, n))) The Wordle info – the same type as the return value from create_wordle_info.

words : A Vector of words of type T.

Return

A subset of the `words` vector based on the filter information
from `wordle_info`.

Examples

julia> (winfo, d) = create_wordle_info("which", "where")

([('w', 1), ('h', 2)], Dict('h' => (0, 0), 'c' => (0, 0), 'i' => (0, 0)))

julia> words = ["state", "which", "where", "child", "there", "taste"]

julia> filter_universe((winfo, d), words)

1-element Vector{T}:
 "where"

Wordle.freq_letter_strat — Function

freq_letter_strat(swords, lfa, c_idx)

Strategy to pick a guess for Wordle.

The strategy consists of the following:

Take the words in the current universe.

Take the complement of the indices where we have exact information. For each of these indices create a dictionary with letter => count.

Pick the index where the corresponding dictionary has the largest count value for some letter.

If for a given dictionary, there are several letters with the same count, pick the letter from the letter freq string below.

Do the same now across dictionaries, find the letter that is the most frequent and its dictionary index.

For this index, find all words with this letter in this slot. Pick the one that is most frequent. This will be our guess.

Type Constraints

T <: AbstractString

Input Contract

The characters in swords are lowercase letters: [a-z].

Arguments

swords::AbstractVector : A Vector of sorted strings (sorted by frequency of occurrence).

lfa::Vector{Char} : This is the alphabet in lower case as a character vector from most to least used.

c_idx::Vector{Int} : This is the index values of words to analyze. This list is usually the complement of exact match indices from a previous guess.

Return

A guess word.

Wordle.get_next_word — Function

get_next_word(words, wts)

This function tries to find the "best" guess from the vector of words, words. It does this by taking each word and rearranging the letters in alphabetical order. It then sorts this list of new "words" lexically, while keeping the associated weights. This list of words is looped over, forming grouping "words" with corresponding associated weight equal to the sum of all proper words that it "contains".

For instance, the words: "ether", "there", and "three" each get translated to the grouping string: "eehrt". Next, the grouping strings like "eehrt" are sorted. In the case of the string "eehrt", it occurs three times for the three associated words. The weights of the three occurrences is also kept. We loop over these grouping strings and add up the associated weights. Finally, we find the grouping string with the most weight and look for the first match in the vector, words, that belongs to this grouping string. Since words is ordered by weight, we get the word that has the highest weight of the words in its associated grouping string.

Type Constraints

T <: AbstractString

Arguments

words::AbstractVector{T} : The remaining pool of words to guess from.

wts ::AbstractVector{Float64} : The usage frequency of the associated words in the vector, words.

Input Contract

words == words[sortperm(wts, rev=true)] $\quad$ (Words are sorted from highest to lowest by word usage.)

Return

The index of a "best" guess in the user supplied vector of words.

Index

Wordle.create_wordle_info

Wordle.filter_universe

Wordle.freq_letter_strat

Wordle.get_next_word

Wordle.solve_wordle